Renamed MicrosoftTelemetry.h to Traceloggingconfig.h and moved the file under wil.
Updated references in Tracelogging.h
