+++ include/wil/traceloggingconfig.h
